- Renamed ConstantNode to ConvertableNode - Moved ConvertableNode into it's own class - Added support for custom conversion code (currently only used by rotation) - Added support for Texture, Normal Map, Cube Texture, Uint, Double, String, Quaternion/Rotation
