+++ Source/Editor/Surface/Archetypes/Constants.cs
@@ -348,96 +348,6 @@ namespace FlaxEditor.Surface.Archetypes
                 SetValue(0, array);
             }
         }
-
-        private class ConstantNode : SurfaceNode
-        {
-            private ScriptType _type;
-            
-            /// <inheritdoc />
-            public ConstantNode(uint id, VisjectSurfaceContext context, NodeArchetype nodeArch, GroupArchetype groupArch, ScriptType type) 
-            : base(id, context, nodeArch, groupArch)
-            {
-                _type = type;
-            }
-
-            /// <inheritdoc />
-            public override void OnShowSecondaryContextMenu(FlaxEditor.GUI.ContextMenu.ContextMenu menu, Float2 location)
-            {
-                base.OnShowSecondaryContextMenu(menu, location);
-                
-                menu.AddSeparator();
-                menu.AddButton("Convert to Parameter", OnConvertToParameter);
-            }
-
-            private void OnConvertToParameter()
-            {
-                if(Surface.Owner is not IVisjectSurfaceWindow window)
-                    throw new Exception("Surface owner is not a Visject Surface Window");
-                
-                Asset asset = Surface.Owner.SurfaceAsset;
-                if (asset == null || !asset.IsLoaded)
-                {
-                    Editor.LogError("Asset is null or not loaded");
-                    return;   
-                }
-
-                // Add parameter to editor
-                var paramIndex = Surface.Parameters.Count;
-                var paramAction = new AddRemoveParamAction
-                {
-                    Window = window,
-                    IsAdd = true,
-                    Name = Utilities.Utils.IncrementNameNumber("New parameter", x => OnParameterRenameValidate(null, x)),
-                    Type = _type,
-                    Index = paramIndex,
-                    InitValue = Values[0],
-                };
-                paramAction.Do();
-
-                var parameterGuid = Surface.Parameters[paramIndex].ID;
-                
-                bool undoEnabled = Surface.Undo.Enabled;
-                Surface.Undo.Enabled = false;
-                SurfaceNode node = Surface.Context.SpawnNode(6, 1, this.Location, new object[] {parameterGuid});
-                Surface.Undo.Enabled = undoEnabled;
-
-                if (node is not Parameters.SurfaceNodeParamsGet getNode)
-                    throw new Exception("Node is not a ParamsGet node!");
-                
-                // Recreate connections of constant node
-                // Constant nodes and parameter nodes should have the same ports, so we can just iterate through the connections
-                var boxes = GetBoxes();
-                for (int i = 0;i < boxes.Count; i++)
-                {
-                    Box box = boxes[i];
-                    if (!box.HasAnyConnection)
-                        continue;
-
-                    if (!getNode.TryGetBox(i, out Box paramBox))
-                        continue;
-                    
-                    // Iterating backwards, because the CreateConnection method deletes entries from the box connections when target box IsSingle is set to true
-                    for (int k = box.Connections.Count-1; k >= 0; k--)
-                    {
-                        Box connectedBox = box.Connections[k];
-                        paramBox.CreateConnection(connectedBox);
-                    }
-                }
-                
-                var spawnNodeAction = new AddRemoveNodeAction(getNode, true);
-                var removeConstantAction = new AddRemoveNodeAction(this, false);
-                
-                Surface.AddBatchedUndoAction(new MultiUndoAction(paramAction, spawnNodeAction, removeConstantAction));
-                removeConstantAction.Do();
-            }
-            
-            private bool OnParameterRenameValidate(RenamePopup popup, string value)
-            {
-                if(Surface.Owner is not IVisjectSurfaceWindow window)
-                    throw new Exception("Surface owner is not a Visject Surface Window");
-                return !string.IsNullOrWhiteSpace(value) && window.VisjectSurface.Parameters.All(x => x.Name != value);
-            }
-        }
         
         /// <summary>
         /// The nodes for that group.
@@ -448,7 +358,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 1,
                 Title = "Bool",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(bool))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(bool))),
                 Description = "Constant boolean value",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(110, 20),
@@ -481,7 +391,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 2,
                 Title = "Integer",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(int))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(int))),
                 Description = "Constant integer value",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(110, 20),
@@ -509,7 +419,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 3,
                 Title = "Float",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(float))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(float))),
                 Description = "Constant floating point",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(110, 20),
@@ -537,7 +447,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 4,
                 Title = "Float2",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(Float2))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(Float2))),
                 Description = "Constant Float2",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(130, 60),
@@ -568,7 +478,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 5,
                 Title = "Float3",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(Float3))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(Float3))),
                 Description = "Constant Float3",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(130, 80),
@@ -601,7 +511,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 6,
                 Title = "Float4",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(Float4))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(Float4))),
                 Description = "Constant Float4",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(130, 100),
@@ -636,7 +546,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 7,
                 Title = "Color",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(Color))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(Color))),
                 Description = "RGBA color",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(70, 100),
@@ -669,6 +579,10 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 8,
                 Title = "Rotation",
+                // TODO: Way too long and ugly - find a better way to add conversion
+                Create = (id, context, arch, groupArch) =>
+                    new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(Quaternion)), 
+                        values => Quaternion.Euler((float)values[0], (float)values[1], (float)values[2])),
                 Description = "Euler angle rotation",
                 Flags = NodeFlags.AnimGraph | NodeFlags.VisualScriptGraph | NodeFlags.ParticleEmitterGraph,
                 Size = new Float2(110, 60),
@@ -693,6 +607,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 9,
                 Title = "String",
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(string))),
                 Description = "Text",
                 Flags = NodeFlags.VisualScriptGraph | NodeFlags.AnimGraph,
                 Size = new Float2(200, 20),
@@ -743,7 +658,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 12,
                 Title = "Unsigned Integer",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(Float3))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(uint))),
                 Description = "Constant unsigned integer value",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(170, 20),
@@ -783,6 +698,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 15,
                 Title = "Double",
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(double))),
                 Description = "Constant floating point",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(110, 20),
@@ -800,7 +716,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 16,
                 Title = "Vector2",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(Vector2))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(Vector2))),
                 Description = "Constant Vector2",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(130, 60),
@@ -821,7 +737,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 17,
                 Title = "Vector3",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(Vector3))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(Vector3))),
                 Description = "Constant Vector3",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(130, 80),
@@ -844,7 +760,7 @@ namespace FlaxEditor.Surface.Archetypes
             {
                 TypeID = 18,
                 Title = "Vector4",
-                Create = (id, context, arch, groupArch) => new ConstantNode(id, context, arch, groupArch, new ScriptType(typeof(Vector4))),
+                Create = (id, context, arch, groupArch) => new ConvertableNode(id, context, arch, groupArch, new ScriptType(typeof(Vector4))),
                 Description = "Constant Vector4",
                 Flags = NodeFlags.AllGraphs,
                 Size = new Float2(130, 100),
